requirements.txt
# Backend framework
fastapi==0.110.0
uvicorn==0.27.1

# Data validation
pydantic==2.6.1

# Core ML & data handling
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.0

# (Optional but safe to add now – used later)
joblib==1.3.2

# Image processing (used in Phase 3 & 4)
opencv-python==4.9.0.80
pillow==10.2.0

# Deep Learning (used later – Phase 3/4)
tensorflow==2.15.0
keras==2.15.0

python-multipart